requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.0
catboost==1.2.2
lightgbm==4.1.0

# Feature Engineering
feature-engine==1.6.2

# MLflow
mlflow==2.8.1
psycopg2-binary==2.9.9

# DVC
dvc==3.30.1
dvc-s3==3.0.1

# AWS
boto3==1.28.85
botocore==1.31.85
awscli==1.29.85

# Model Serving - Triton
tritonclient[all]==2.38.0

# Ray Serve
ray[serve]==2.8.0
ray[default]==2.8.0

# LLM Libraries
langchain==0.0.350
langchain-community==0.0.1
llama-index==0.9.14
sentence-transformers==2.2.2
transformers==4.35.2
accelerate==0.25.0
bitsandbytes==0.41.3

# Vector Databases
qdrant-client==1.7.0
pinecone-client==2.2.4

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
locust==2.17.0
great-expectations==0.18.5

# Model Conversion
onnx==1.15.0
onnxruntime==1.16.3
tf2onnx==1.15.1

# Monitoring
prometheus-client==0.19.0
grafana-api==1.0.3

# Data Processing
pyarrow==14.0.1
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
requests==2.31.0

# Logging
loguru==0.7.2

# Notebook
jupyter==1.0.0
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0